refactor: move enums to defintions modules
