+++ ohsome_quality_analyst/topics/definitions.py
@@ -1,0 +1,5 @@
+from enum import Enum
+
+from ohsome_quality_analyst.definitions import get_topic_keys
+
+TopicEnum = Enum("TopicEnum", {name: name for name in get_topic_keys()})
